backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
psycopg2-binary==2.9.6
aiofiles==23.1.0
aiohttp==3.8.4
bcrypt==4.0.1